requirements.txt
flask
flask-socketio
python-dotenv
openai
pyOpenSSL
websockets
gunicorn
gevent
gevent-websocket
